Re-did the hotkeys for all.
You can now enable / disable modifers and get feedback if the current key is enabled / disabled.
Raise Ally is now on focus target only. You can bind SetFocus on a Key inside KeyBindings on WoW, so you can push that key to set focus then revive ally during combat.
AMZ is still the same, when you push down the hotkey for enable it will still use AMZ on your current mouse cursor.
Army of the Dead key is there if you want to save it for later(Note, you can leave this hotkey enabled, if you want it to use Army of the Dead when it can during the encounter)
Tier6 abilities key is still the same, it will use the tier 6 ability you have for level 90 talent(When starting up the bot).